# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lang Cards API v1",
        "description": "Lang Cards API ",
        "contact": {
            "email": "vlad2000100600@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/sendVerificationCodeEmail": {
            "post": {
                "tags": [
                    "Верификация электронного адреса"
                ],
                "summary": "Отправка кода подтверждения на email",
                "description": "Отправляет код подтверждения на email текущего авторизованного пользователя, если его email ещё не был подтверждён.",
                "operationId": "sendVerificationCodeEmail",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptLanguageHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Код подтверждения успешно отправлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Сообщение с кодом для подтверждения email - адреса было отправлено на электронный адрес, указанный при регистрации"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Email уже подтверждён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email - адрес текущего авторизованного пользователя уже был ранее подтвержден!"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "description": "Ошибка при генерации кода",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка при генерации кода"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/verificationEmailAddress": {
            "post": {
                "tags": [
                    "Верификация электронного адреса"
                ],
                "summary": "Подтверждение email авторизованного пользователя кодом",
                "description": "Позволяет подтвердить email пользователя по коду, отправленному на электронную почту.",
                "operationId": "verificationEmailAddress",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptLanguageHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailVerificationCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email успешно подтверждён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Электронный адрес авторизованного пользователя был успешно подтвержден"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь с id = 5 не найден"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Email уже подтверждён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email - адрес текущего авторизованного пользователя уже был ранее подтвержден!"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Код не был запрошен ранее",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Текущий авторизованный пользователь не запрашивал код для подтверждения почты"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Неверный код подтверждения",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Предоставленный код не соответствует коду из электронного сообщения"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Предоставленные данные не валидны"
                                        },
                                        "errors": {
                                            "properties": {
                                                "code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Token is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmailVerificationCodeRequest": {
                "title": "Email Verification Code Request (Данные для подтверждения электронной почты)",
                "description": "Запрос на подтверждение email кодом",
                "required": [
                    "code"
                ],
                "properties": {
                    "code": {
                        "description": "Код подтверждения email, состоящий из ровно 6 цифр",
                        "type": "string",
                        "pattern": "^\\d{6}$",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthorized": {
                "description": "Пользователь не авторизован",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Пользователь не авторизован и не имеет доступа к данным"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": null,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "AcceptLanguageHeader": {
                "name": "Accept-Language",
                "in": "header",
                "description": "Язык локали",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "ru",
                    "enum": [
                        "ru",
                        "en"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}